combine emitter tests
